phase2_ai_model/requirements.txt
python-dotenv==1.0.0
openai==1.10.0
anthropic==0.7.0
langchain==0.0.320
langchain-openai==0.0.5
pydantic==2.5.0
pyyaml==6.0
requests==2.31.0
pytest==7.4.3
black==23.12.0
jsonschema==4.20.0
